+++ tests/Normalizer/SimpleNormalizerTest.php
@@ -218,6 +218,46 @@ final class SimpleNormalizerTest extends TestCase
 	}
 
 	/**
+	 */
+	public function testWithEntityManagerCachesNormalizedClassName () : void
+	{
+		$classMetaData = new ClassMetadata("SomeClass");
+
+		$metadataFactory = $this->createMock(ClassMetadataFactory::class);
+		$metadataFactory
+			->expects(self::once())
+			->method("hasMetadataFor")
+			->with(DummyVO::class)
+			->willReturn(true);
+
+		$metadataFactory
+			->expects(self::once())
+			->method("getMetadataFor")
+			->with(DummyVO::class)
+			->willReturn($classMetaData);
+
+		$entityManager = $this->createMock(EntityManagerInterface::class);
+		$entityManager->method("getMetadataFactory")->willReturn($metadataFactory);
+
+		$locator = $this->createMock(ServiceLocator::class);
+
+		$locator->expects(self::exactly(2))
+			->method("get")
+			->with("SomeClass")
+			->willReturn(new DummyVONormalizer(5));
+
+		$normalizer = new SimpleNormalizer(
+			objectNormalizers: $locator,
+			isDebug: true,
+			validJsonVerifier: new ValidJsonVerifier(),
+			entityManager: $entityManager,
+		);
+
+		$normalizer->normalize(new DummyVO(5));
+		$normalizer->normalize(new DummyVO(5));
+	}
+
+	/**
 	 * @return ServiceLocator<mixed>
 	 */
 	private function createNormalizerObjectNormalizers (mixed $returnValue) : ServiceLocator